# openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LangPlug API",
    "description": "German Language Learning Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/videos": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get Videos",
        "description": "Get list of available videos/series",
        "operationId": "get_videos_api_videos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VideoInfo"
                  },
                  "type": "array",
                  "title": "Response Get Videos Api Videos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/videos/subtitles/{subtitle_path}": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get Subtitles",
        "description": "Serve subtitle files - Requires authentication",
        "operationId": "get_subtitles_api_videos_subtitles__subtitle_path__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subtitle_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subtitle Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{series}/{episode}": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Stream Video",
        "description": "Stream video file - Requires authentication",
        "operationId": "stream_video_api_videos__series___episode__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "series",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series"
            }
          },
          {
            "name": "episode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/subtitle/upload": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Upload Subtitle",
        "description": "Upload subtitle file for a video - Requires authentication",
        "operationId": "upload_subtitle_api_videos_subtitle_upload_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "video_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_subtitle_api_videos_subtitle_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/scan": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Scan Videos",
        "description": "Scan videos directory for new videos - Requires authentication",
        "operationId": "scan_videos_api_videos_scan_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/videos/user": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get User Videos",
        "description": "Get videos for the current user (alias for get_available_videos)",
        "operationId": "get_user_videos_api_videos_user_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get User Videos Api Videos User Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/videos/{video_id}/vocabulary": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get Video Vocabulary",
        "description": "Get vocabulary words extracted from a video",
        "operationId": "get_video_vocabulary_api_videos__video_id__vocabulary_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VocabularyWord"
                  },
                  "title": "Response Get Video Vocabulary Api Videos  Video Id  Vocabulary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}/status": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get Video Status",
        "description": "Get processing status for a video",
        "operationId": "get_video_status_api_videos__video_id__status_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/upload": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Upload Video Generic",
        "description": "Upload a new video file (generic endpoint) - Requires authentication",
        "operationId": "upload_video_generic_api_videos_upload_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "series",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Default",
              "title": "Series"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_generic_api_videos_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/upload/{series}": {
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "Upload Video To Series",
        "description": "Upload a new video file to a series - Requires authentication",
        "operationId": "upload_video_to_series_api_videos_upload__series__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "series",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_to_series_api_videos_upload__series__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_subtitle_api_videos_subtitle_upload_post": {
        "properties": {
          "subtitle_file": {
            "type": "string",
            "format": "binary",
            "title": "Subtitle File"
          }
        },
        "type": "object",
        "required": [
          "subtitle_file"
        ],
        "title": "Body_upload_subtitle_api_videos_subtitle_upload_post"
      },
      "Body_upload_video_generic_api_videos_upload_post": {
        "properties": {
          "video_file": {
            "type": "string",
            "format": "binary",
            "title": "Video File"
          }
        },
        "type": "object",
        "required": [
          "video_file"
        ],
        "title": "Body_upload_video_generic_api_videos_upload_post"
      },
      "Body_upload_video_to_series_api_videos_upload__series__post": {
        "properties": {
          "video_file": {
            "type": "string",
            "format": "binary",
            "title": "Video File"
          }
        },
        "type": "object",
        "required": [
          "video_file"
        ],
        "title": "Body_upload_video_to_series_api_videos_upload__series__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessingStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "subtitle_path": {
            "type": "string",
            "title": "Subtitle Path",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "status",
          "progress",
          "current_step"
        ],
        "title": "ProcessingStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoInfo": {
        "properties": {
          "series": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Series",
            "description": "Name of the TV series"
          },
          "season": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Season",
            "description": "Season identifier (e.g., 'S01', 'Season 1')"
          },
          "episode": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Episode",
            "description": "Episode identifier (e.g., 'E01', 'Episode 1')"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "Episode title"
          },
          "path": {
            "type": "string",
            "minLength": 1,
            "title": "Path",
            "description": "File system path to the video file"
          },
          "has_subtitles": {
            "type": "boolean",
            "title": "Has Subtitles",
            "description": "Whether subtitles are available for this video"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration",
            "description": "Video duration in seconds"
          }
        },
        "type": "object",
        "required": [
          "series",
          "season",
          "episode",
          "title",
          "path",
          "has_subtitles"
        ],
        "title": "VideoInfo",
        "example": {
          "duration": 1320.5,
          "episode": "E01",
          "has_subtitles": true,
          "path": "/videos/Superstore/S01/E01.mp4",
          "season": "S01",
          "series": "Superstore",
          "title": "Pilot"
        }
      },
...
      "VocabularyWord": {
        "properties": {
          "word": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Word",
            "description": "The vocabulary word"
          },
          "definition": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition",
            "description": "Definition of the word"
          },
          "difficulty_level": {
            "type": "string",
            "pattern": "^(A1|A2|B1|B2|C1|C2)$",
            "title": "Difficulty Level",
            "description": "CEFR difficulty level (A1, A2, B1, B2, C1, C2)"
          },
          "known": {
            "type": "boolean",
            "title": "Known",
            "description": "Whether the user knows this word",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "word",
          "difficulty_level"
        ],
        "title": "VocabularyWord"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}